Switched to use standard Avatar component.
