+++ frontend/src/components/UserInfo/styles.module.css
@@ -25,8 +25,6 @@
     position: absolute;
     top: 0;
     left: 0;
-    width: 200px;
-    height: 200px;
 }
 
 .userDetailsContainer {
@@ -46,15 +44,6 @@
     grid-column: 1 / -1;
     margin-top: 20px;
     overflow-y: auto;
-}
-
-.userInfoAvatar img {
-    aspect-ratio: 1;
-    width: 100%; /* Ensures the image fills the container */
-    border-radius: 50%;
-    object-fit: cover;
-    box-shadow: 0 0 10px #fff9e6;
-    overflow: hidden;
 }
 
 .userInfoFormSocial {
